requirements.txt
langchain
openai
python-dotenv
presidio-analyzer
presidio-anonymizer
langchain-experimental
spacy
Faker
fasttext-wheel
openpyxl
matplotlib